+++ diesel_cli/src/config.rs
@@ -3,7 +3,7 @@ use std::env;
 use std::error::Error;
 use std::fs;
 use std::io::Read;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 use super::find_project_root;
 use crate::print_schema;
@@ -31,11 +31,17 @@ impl Config {
 
         if path.exists() {
             let mut bytes = Vec::new();
-            fs::File::open(path)?.read_to_end(&mut bytes)?;
-            toml::from_slice(&bytes).map_err(Into::into)
+            fs::File::open(&path)?.read_to_end(&mut bytes)?;
+            let mut result = toml::from_slice::<Self>(&bytes)?;
+            result.set_relative_path_base(path.parent().unwrap());
+            Ok(result)
         } else {
             Ok(Self::default())
         }
+    }
+
+    fn set_relative_path_base(&mut self, base: &Path) {
+        self.print_schema.set_relative_path_base(base)
     }
 }
 
@@ -63,6 +69,20 @@ impl PrintSchema {
 
     pub fn import_types(&self) -> Option<&[String]> {
         self.import_types.as_ref().map(|v| &**v)
+    }
+
+    fn set_relative_path_base(&mut self, base: &Path) {
+        if let Some(ref mut file) = self.file {
+            if file.is_relative() {
+                *file = base.join(&file);
+            }
+        }
+
+        if let Some(ref mut patch_file) = self.patch_file {
+            if patch_file.is_relative() {
+                *patch_file = base.join(&patch_file);
+            }
+        }
     }
 }