+++ diesel_cli/src/database_error.rs
@@ -10,7 +10,7 @@ pub type DatabaseResult<T> = Result<T, DatabaseError>;
 
 #[derive(Debug)]
 pub enum DatabaseError {
-    CargoTomlNotFound,
+    ProjectRootNotFound,
     DatabaseUrlMissing,
     IoError(io::Error),
     QueryError(result::Error),
@@ -40,8 +40,8 @@ impl Error for DatabaseError {}
 impl fmt::Display for DatabaseError {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         match *self {
-            CargoTomlNotFound => {
-                f.write_str("Unable to find Cargo.toml in this directory or any parent directories.")
+            ProjectRootNotFound => {
+                f.write_str("Unable to find diesel.toml or Cargo.toml in this directory or any parent directories.")
             }
             DatabaseUrlMissing => {
                 f.write_str("The --database-url argument must be passed, or the DATABASE_URL environment variable must be set.")
@@ -65,7 +65,7 @@ impl fmt::Display for DatabaseError {
 impl PartialEq for DatabaseError {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
-            (&CargoTomlNotFound, &CargoTomlNotFound) => true,
+            (&ProjectRootNotFound, &ProjectRootNotFound) => true,
             _ => false,
         }
     }