+++ diesel_cli/src/main.rs
@@ -293,7 +293,7 @@ fn generate_completions_command(matches: &ArgMatches) {
 /// Looks for a migrations directory in the current path and all parent paths,
 /// and creates one in the same directory as the Cargo.toml if it can't find
 /// one. It also sticks a .gitkeep in the directory so git will pick it up.
-/// Returns a `DatabaseError::CargoTomlNotFound` if no Cargo.toml is found.
+/// Returns a `DatabaseError::ProjectRootNotFound` if no Cargo.toml is found.
 fn create_migrations_directory(path: &Path) -> DatabaseResult<PathBuf> {
     println!("Creating migrations directory at: {}", path.display());
     fs::create_dir(path)?;
@@ -302,19 +302,21 @@ fn create_migrations_directory(path: &Path) -> DatabaseResult<PathBuf> {
 }
 
 fn find_project_root() -> DatabaseResult<PathBuf> {
-    search_for_cargo_toml_directory(&env::current_dir()?)
+    let current_dir = env::current_dir()?;
+    search_for_directory_containing_file(&current_dir, "diesel.toml")
+        .or_else(|_| search_for_directory_containing_file(&current_dir, "Cargo.toml"))
 }
 
 /// Searches for the directory that holds the project's Cargo.toml, and returns
-/// the path if it found it, or returns a `DatabaseError::CargoTomlNotFound`.
-fn search_for_cargo_toml_directory(path: &Path) -> DatabaseResult<PathBuf> {
-    let toml_path = path.join("Cargo.toml");
+/// the path if it found it, or returns a `DatabaseError::ProjectRootNotFound`.
+fn search_for_directory_containing_file(path: &Path, file: &str) -> DatabaseResult<PathBuf> {
+    let toml_path = path.join(file);
     if toml_path.is_file() {
         Ok(path.to_owned())
     } else {
         path.parent()
-            .map(search_for_cargo_toml_directory)
-            .unwrap_or(Err(DatabaseError::CargoTomlNotFound))
+            .map(|p| search_for_directory_containing_file(p, file))
+            .unwrap_or(Err(DatabaseError::ProjectRootNotFound))
     }
 }
 
@@ -475,7 +477,7 @@ mod tests {
     use std::path::PathBuf;
 
     use super::convert_absolute_path_to_relative;
-    use super::search_for_cargo_toml_directory;
+    use super::search_for_directory_containing_file;
 
     #[test]
     fn toml_directory_find_cargo_toml() {
@@ -487,7 +489,7 @@ mod tests {
 
         assert_eq!(
             Ok(temp_path.clone()),
-            search_for_cargo_toml_directory(&temp_path)
+            search_for_directory_containing_file(&temp_path, "Cargo.toml")
         );
     }
 
@@ -497,8 +499,8 @@ mod tests {
         let temp_path = dir.path().canonicalize().unwrap();
 
         assert_eq!(
-            Err(DatabaseError::CargoTomlNotFound),
-            search_for_cargo_toml_directory(&temp_path)
+            Err(DatabaseError::ProjectRootNotFound),
+            search_for_directory_containing_file(&temp_path, "Cargo.toml")
         );
     }